feat: create TasksWorkManager
+ start work when user is authorized
+ cancel work when user signs out
